pyproject.toml
[tool.poetry]
name = "final-work"
version = "0.1.0"
description = ""
authors = ["TatianaBelova333 <101121235+TatianaBelova333@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "final_work"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.14.0"
python-multipart = "^0.0.5"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
marshmallow-dataclass = "^8.5.9"
requests = "^2.28.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
